Addressing code review comments for ETL-3251
